Fix for ICE-7342 (Improve ice:dataExporter memory usage) to icefaces 2

+++ icefaces2/trunk/icefaces/compat/components/src/main/java/com/icesoft/faces/component/dataexporter/OutputTypeHandler.java
@@ -34,12 +34,14 @@ public abstract class OutputTypeHandler {
 	}
 
 	public OutputTypeHandler(String path) {
-		try {
-			file = new File(path);
-			file.createNewFile();
-			file.deleteOnExit();
-		} catch (IOException ioe) {
-			ioe.printStackTrace();
+		if (!"no-data".equals(path)) {
+			try {
+				file = new File(path);
+				file.createNewFile();
+				file.deleteOnExit();
+			} catch (IOException ioe) {
+				ioe.printStackTrace();
+			}
 		}
 	}
